+++ ci4/app/Interfaces/HttpProxyRoute.php
@@ -3,6 +3,9 @@
  * @property string $path
  * @property int $port
  * @property string $protocol
+ * @property string $timeoutPolicyIdle
+ * @property string $timeoutPolicyResponse
+ * @property string $timeoutPolicyIdleConnection
  */
 interface HttpProxyRoute {