+++ vue/src/components/Dialogs/Setup/DeploymentSpecifications/UpdateDialogs/DeploymentSpecificationUpdateHttpProxyRouteDialog.vue
@@ -7,6 +7,9 @@ export interface DeploymentSpecificationUpdateHttpProxyRoutePathDialog_Input {
         path: string;
         port: number;
         protocol: string;
+        timeoutPolicyIdle: string;
+        timeoutPolicyResponse: string;
+        timeoutPolicyIdleConnection: string;
     };
 
     onSaveCallback: () => void;
@@ -27,6 +30,9 @@ const protocolItems = ref([
     'tls', 'h2', 'h2c',
 ]);
 const protocol = ref('');
+const timeoutPolicyIdle = ref('');
+const timeoutPolicyResponse = ref('');
+const timeoutPolicyIdleConnection = ref('');
 
 const rules = {
     required: [
@@ -56,6 +62,9 @@ function render() {
     path.value = props.input.httpProxyRoute.path ?? '';
     port.value = props.input.httpProxyRoute.port ?? '';
     protocol.value = props.input.httpProxyRoute.protocol ?? '';
+    timeoutPolicyIdle.value = props.input.httpProxyRoute.timeoutPolicyIdle ?? '';
+    timeoutPolicyResponse.value = props.input.httpProxyRoute.timeoutPolicyResponse ?? '';
+    timeoutPolicyIdleConnection.value = props.input.httpProxyRoute.timeoutPolicyIdleConnection ?? '';
     showDialog.value = true;
 }
 
@@ -72,6 +81,9 @@ function onSaveBtnClicked() {
     props.input.httpProxyRoute.path = path.value;
     props.input.httpProxyRoute.port = port.value ?? 0;
     props.input.httpProxyRoute.protocol = protocol.value;
+    props.input.httpProxyRoute.timeoutPolicyIdle = timeoutPolicyIdle.value;
+    props.input.httpProxyRoute.timeoutPolicyResponse = timeoutPolicyResponse.value;
+    props.input.httpProxyRoute.timeoutPolicyIdleConnection = timeoutPolicyIdleConnection.value;
     props.input.onSaveCallback();
     close();
 }
@@ -125,6 +137,36 @@ function onCloseBtnClicked() {
                                 :items="protocolItems"
                                 clearable
                                 hint="omit to fall back on service annotations"
+                                persistent-hint
+                            />
+                        </v-col>
+                        <v-col cols="6">
+                            <v-text-field
+                                v-model="timeoutPolicyIdle"
+                                variant="outlined"
+                                label="Timeout Policy: Idle"
+                                clearable
+                                hint="Timeout for receiving a response from the server after processing a request from client. If not supplied, Envoy’s default value of 15s applies. More information can be found in Envoy’s documentation."
+                                persistent-hint
+                            />
+                        </v-col>
+                        <v-col cols="6">
+                            <v-text-field
+                                v-model="timeoutPolicyResponse"
+                                variant="outlined"
+                                label="Timeout Policy: Response"
+                                clearable
+                                hint="Timeout for how long the proxy should wait while there is no activity during single request/response (for HTTP/1.1) or stream (for HTTP/2). Timeout will not trigger while HTTP/1.1 connection is idle between two consecutive requests. If not specified, there is no per-route idle timeout, though a connection manager-wide stream idle timeout default of 5m still applies. More information can be found in Envoy’s documentation."
+                                persistent-hint
+                            />
+                        </v-col>
+                        <v-col cols="6">
+                            <v-text-field
+                                v-model="timeoutPolicyIdleConnection"
+                                variant="outlined"
+                                label="Timeout Policy: Idle Connection"
+                                clearable
+                                hint="Timeout for how long connection from the proxy to the upstream service is kept when there are no active requests. If not supplied, Envoy’s default value of 1h applies. More information can be found in Envoy’s documentation."
                                 persistent-hint
                             />
                         </v-col>